requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
docstring-parser==0.16
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.4.0
google-auth==2.40.1
google-auth-oauthlib>0.5,<2
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
identify==2.6.10
idna==3.10
imageio==2.37.0
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
llvmlite==0.44.0
magicgui==0.10.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
ml-dtypes==0.3.2
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
pandas==2.2.3
patchify==0.2.3
pillow==11.2.1
platformdirs==4.3.8
pre-commit==4.2.0
protobuf>=3.20.3,<5.0.0dev
psygnal==0.13.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyconify==0.2.1
pydantic==2.11.4
pydantic-core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
QtPy==2.4.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.15.3
six==1.17.0
skan==0.12.2
sniffio==1.3.1
starlette==0.46.2
superqt==0.7.3
tensorboard==2.15.1
tensorboard-data-server>=0.7.0,<0.8.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.4.0
tifffile==2025.5.10
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
Werkzeug==3.1.3
wrapt>=1.11.0,<1.15
